Update linux tcp client.cpp

+++ linux tcp client.cpp
@@ -31,7 +31,7 @@ void* thread_read(void* arg) {
     while (1) {
         usleep(1);
         memset(buf, 0, sizeof(buf));
-        size_t size = read(sd, buf, sizeof(buf));
+        int size = read(sd, buf, sizeof(buf));
         if (size <= 0) {
             continue;
         }
